requirements.txt
python-dotenv
rich
pyfiglet
click
tqdm
pydantic
